requirements.txt
langchain
langgraph
langchain-core
langchain-community
ipykernel
python-dotenv
langchain-groq
duckduckgo-search
arxiv
wikipedia
langchain_mcp_adapters
mcp
uvicorn[standard]
pymupdf 
pydantic